chore(web): splits banner.ts into separate files per banner type
