+++ web/src/engine/osk/src/banner/bannerView.ts
@@ -1,12 +1,14 @@
 import EventEmitter from 'eventemitter3';
 
-import { Banner, BlankBanner, ImageBanner, SuggestionBanner } from './banner.js';
-import OSKViewComponent from '../components/oskViewComponent.interface.js';
-import { ParsedLengthStyle } from '../lengthStyle.js';
-
 import { DeviceSpec } from '@keymanapp/web-utils';
 import type { PredictionContext, StateChangeEnum } from '@keymanapp/input-processor';
 import { createUnselectableElement } from 'keyman/engine/dom-utils';
+
+import { Banner, BlankBanner } from './banner.js';
+import { ImageBanner } from './imageBanner.js';
+import OSKViewComponent from '../components/oskViewComponent.interface.js';
+import { ParsedLengthStyle } from '../lengthStyle.js';
+import { SuggestionBanner } from './suggestionBanner.js';
 
 /**
  * This object is used to specify options by both `BannerManager.getOptions`
